Add files via upload
Pretty close 1 to 1 of names to IDs from different sources. Underdog does not provide easy matching here so I took some time to map these.

I've done this in a more manual and painstaking way in the past. This file should hasten the process to join in other data sources to their publically available data